requirements.txt
arabic-reshaper
python-bidi
arabicscript